# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/beautician": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "美容師情報取得",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel.BeauticianGet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsemodel.BeauticianGet"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/resource.Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "美容師情報更新",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel.BeauticianUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/resource.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "美容師登録",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel.BeauticianCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsemodel.BeauticianCreate"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/resource.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/beautician/find": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "美容師検索",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel.BeauticianFind"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsemodel.BeauticianFind"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/resource.Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/reservation": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "予約登録",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel.ReservationCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsemodel.ReservationCreate"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/resource.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/beautician": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "美容師予約情報取得",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel.ReservationFindByBeautician"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsemodel.ReservationFindByBeautician"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/resource.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/find": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "予約検索",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel.ReservationFind"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsemodel.ReservationFind"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/resource.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/user": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ゲストの予約履歴を取得",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel.ReservationFindByUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsemodel.ReservationFindByUser"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/resource.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/reservation/{randID}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "予約詳細を取得",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel.ReservationGetInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsemodel.ReservationGetInfo"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/resource.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "requestmodel.BeauticianCreate": {
            "type": "object",
            "properties": {
                "instagramId": {
                    "type": "string"
                },
                "lineId": {
                    "type": "string"
                }
            }
        },
        "requestmodel.BeauticianFind": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "menuRandIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "salonRandID": {
                    "type": "string"
                }
            }
        },
        "requestmodel.BeauticianGet": {
            "type": "object"
        },
        "requestmodel.BeauticianUpdate": {
            "type": "object",
            "required": [
                "firstName",
                "firstNameKana",
                "lastName",
                "lastNameKana",
                "phoneNumber"
            ],
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "firstNameKana": {
                    "type": "string"
                },
                "instagramId": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "lastNameKana": {
                    "type": "string"
                },
                "lineId": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
...
        "requestmodel.ReservationCreate": {
            "type": "object",
            "properties": {
                "beauticianRandId": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "menuIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "salonRandId": {
                    "type": "string"
                }
            }
        },
        "requestmodel.ReservationFind": {
            "type": "object",
            "properties": {
                "beauticianRandID": {
                    "type": "string"
                }
            }
        },
        "requestmodel.ReservationFindByBeautician": {
            "type": "object",
            "properties": {
                "offset": {
                    "type": "integer"
                }
            }
        },
        "requestmodel.ReservationFindByUser": {
            "type": "object"
        },
        "requestmodel.ReservationGetInfo": {
            "type": "object",
            "properties": {
                "randId": {
                    "type": "string"
                }
            }
        },
...
        "resource.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responsemodel.Beautician": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "instagramId": {
                    "type": "string"
                },
                "lineId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "responsemodel.BeauticianCreate": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "instagramId": {
                    "type": "string"
                },
                "lineId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "responsemodel.BeauticianFind": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodel.User"
                    }
                }
            }
        },
        "responsemodel.BeauticianGet": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "instagramId": {
                    "type": "string"
                },
                "lineId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "responsemodel.BeauticianMenu": {
            "type": "object",
            "properties": {
                "beauticianId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "menuId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "responsemodel.Reservation": {
            "type": "object",
            "properties": {
                "beauticiaId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "menuId": {
                    "type": "integer"
                },
                "randId": {
                    "type": "string"
                },
                "spaceId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "responsemodel.ReservationCreate": {
            "type": "object",
            "properties": {
                "beauticiaId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "menuId": {
                    "type": "integer"
                },
                "randId": {
                    "type": "string"
                },
                "spaceId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "responsemodel.ReservationFind": {
            "type": "object",
            "properties": {
                "reservations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodel.Reservation"
                    }
                }
            }
        },
        "responsemodel.ReservationFindByBeautician": {
            "type": "object",
            "properties": {
                "reservations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodel.Reservation"
                    }
                }
            }
        },
        "responsemodel.ReservationFindByUser": {
            "type": "object",
            "properties": {
                "reservations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodel.ReservationGetByUser"
                    }
                }
            }
        },
        "responsemodel.ReservationGetByUser": {
            "type": "object",
            "properties": {
                "beauticianFirstName": {
                    "type": "string"
                },
                "beauticianLastName": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodel.BeauticianMenu"
                    }
                },
                "salonName": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "responsemodel.ReservationGetInfo": {
            "type": "object",
            "properties": {
                "reservationInfo": {
                    "$ref": "#/definitions/responsemodel.ReservationInfo"
                }
            }
        },
        "responsemodel.ReservationInfo": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodel.BeauticianMenu"
                    }
                },
                "randId": {
                    "type": "string"
                },
                "salon": {
                    "$ref": "#/definitions/responsemodel.Salon"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/responsemodel.User"
                }
            }
        },
        "responsemodel.Salon": {
            "type": "object",
            "properties": {
                "addressCode": {
                    "type": "string"
                },
                "addressOther": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "closingHours": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "openingHours": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                },
                "prefectures": {
                    "type": "string"
                },
                "randId": {
                    "type": "string"
                },
                "town": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "responsemodel.User": {
            "type": "object",
            "properties": {
                "beauticianInfo": {
                    "$ref": "#/definitions/responsemodel.Beautician"
                },
                "beauticianMenus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodel.BeauticianMenu"
                    }
                },
                "beauticianSalons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsemodel.Salon"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "firstNameKana": {
                    "type": "string"
                },
                "isBeautician": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "lastNameKana": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "randId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}